feat(card-builder): add home screen with saved cards
